Updating ordering for rownr
Realised that I had taken the row numbering of the file as canon, and that turned out to not be the case - it's based on uploaded time. SO reordering, such that rownr will now reflect delivery time, which will matter for anything depending on lag (i.e. MSSD, emotion half-life).